requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-experimental
python-dotenv
SQLAlchemy